Changed again the way LLVM module is identified for _LLVMModuleUtils.__string_constants
Originally, string constants were identified by the tuple (module, const_str)
But in Python 3, LLVMModule is not hashable and therefore cannot be used as
a dictionary key. The first try was using (module.id, const_str), but this
fails in certain test cases. This patch uses (id(module), const_str), which
is suboptimal as id() is not a real hash. In the long run, it will be nice
to have implement __hash__ in LLVMModule.

90 Tests in total. (Skip: 1. Theano not installed)
Python 2.7:  1/90 FAILED
Python 3.2:  9/90 FAILED