Rebased Impala work off of devel and fixed some bugs
The Impala branch was getting quite diverged, so I manually moved the
work into a new branch off of devel, since it's mostly orthogonal.
Attempting and actual rebase/merge was complicated because of a nonlinear
history that made some features come in and out.

Also fixed some whitespace issues and a bug regarding return types in the
ABI Handling.