types.int_ is now mapped to numpy.intc
Previously, types.int_ was mapped to numpy.int_, which the numpy docs
state is mapped to a C long.  This is probably a mistake, and it causes
ABI errors when I try to map an external C signature that returns int
which was erroneously lowered to int64 on my machine instead of int32.