Forcibly get a writable buffer for struct types
NumPy structured types only provide read-only buffer. However, we
need to write to the buffer in the case of records. This is
already achieved for the CPU backend by simply writing to the
read-only buffer directly.

However, the CUDA backend requires a writable buffer. In order to
get a writable buffer for structured types, we get a read-only
buffer then construct a new read-write buffer pointing to the same
location.