Fix #1734: NRT decref bug when variable is del'ed before being defined
With the new del insertion code, a variable can be del'ed before it is defined
in IR lexical order.  This is a perfectly valid situation, for example a
variable can be del'ed at the beginning of a loop's body before being defined
later in the loop, if the control flow graph requires to do so.  However, the
lowering logic failed to actually emit the decref in such cases.