Use system hypot/hypotf where possible.
This is an attempt at fixing the known overflow problems in the
`hypot()` impl in numba by delegating all hypot calls to the
local system maths library. For platforms with C99 both float
and double variants exist, for those which don't have the float
variant (some windows) floats are cast to doubles and the double
implementation used, with a cast back to float on return.

This also fixes a bug in the hypot unit test which meant it was
not running the no-python-mode variant of the test.

Relevant issues/PRs:
 * https://github.com/numba/numba/issues/1877
 * https://github.com/numba/numba/pull/1619
 * https://github.com/numba/numba/issues/1570