Force homogeneous dtypes in linalg solve/lstsq.
This fixes #1949. After discussion it was decided to deviate from
numpy behavior whereby type promotion occurs to match the widest
dtype in the `linalg.{solve,lstsq}` functions. In numba the types
of the args must be homogenous, this is synonymous with performance
as no inadvertent copying takes place to handle promotion. Users
can trivially implement such bevahviour if they wish outside of the
function call sites.