Refactor LAPACK/BLAS signatures.
This PR merely aims to neaten the call sites where a LAPACK or
BLAS function is required, no external functional changes are
made. The LAPACK/BLAS type signatures and bindings are used in
multiple places in the `np.linalg` module and so it makes sense
for their definitions to be done upfront and only once, this also
reduces the SLOC required for each function making a call to one
of these libraries.