Fix segfault in Appveyor build with Python 3.5
The appveyor build kept segfaulting during the build.cmd script
in the command:
```
python -m pip install -e .
```
Doing the same with a debugger available showed the segfault was
inconsistent but in `os_spawnve` when it occurred.

In testing, it was found Python < 3.5 does not have this problem.
@sklam noted https://github.com/numpy/numpy/issues/7607
and experienced the same issue (many thanks).

As a fix to get Appveyor working again, this patch moves the
test Python version to 3.4.