avoid dead code remove for aliased linalg vars
Linear algebra functions in C (lapack etc) are called
with aliased inputs. Liveness ir.Del insertion and dead code elimination
shouldn't remove those inputs. Previously, an attribute was referenced
for liveness to work but this isn't enough for dead code elimination.
This patch uses a dummy function for this purpose.