Make the dependency on compilers for AOT optional.
This patch removes the dependency on the presence of the Anaconda
compiler toolchain at runtime which is needed for AOT compilation.

If AOT is attempted without the compilers present a RuntimeError is
raised with a message about obtaining the right toolchain. Tests
are updated to skip AOT testing if the right compilers are not
present, conda recipes now have compilers purely as a test
dependency, docs are updated.