Fix tests to be resiliant to np failures.
If an array with a zero dimension is passed to certain NumPy
linalg functions using NumPy < 1.13 then spurious errors are
raised. Post 1.13 defined behaviours were added. This patch
adds code to skip the checking code if a bad combination is
hit, but it still runs the Numba implementation to make sure
it doesn't raise under the same conditions.