Fix random permutation test.
This fixes the random permutation test. It was sporadically failing
on Python 2.7 as occasionally a permutation was picked that was the
identity permutation and the test was asserting that:

```
input != permutation(input)
```

most of the time as input was large and linear this wasn't a problem
but occasionally for permutations in the first axis on arrays with a
small first dimension the identity permutation would appear and cause
the test to fail.