Remove old recipes, switch to test config based on env var.
This removes some old conda recipes that were useful at a time
prior to conda-build baking all the config into the pkg at build
time. It now allows for a single `TESTS_TO_RUN` env var to be set
that determines the subset of tests to run, if the env var is not
set it defaults to all tests.