Fixes issue with missing pygments in testing and adds guards.
This fixes an issue seen on .e.g ARMv8 where there is no pygments
package available. Tests were failing as there was no guard to
prevent them running and the actual code had no check to make sure
that the dependency was present in the case it wasn't.

This also adds a `VANILLA_INSTALL` setting to the Travis CI builds
which is a Numba test suite evaluation in the presence of none of
the extra packages being available. This is to catch tests that
exercise code paths relying on packages being present that do not
have suitable guard mechanisms.