Add `long_running` test flag and feature to exclude tests.
There are certain tests which necessarily take a very long time
as they tax CPU resources. It is not appropriate for these tests
to run on public CI as it they contribute significantly to testing
times and sometimes cause timeouts on CI infrastructure. This
patch adds the test tag `@tag('long_test')` and the option
`--exclude-tags` to the test suite runner. Using these in
combination along with marking known long tests should speed up
test times in CI.