Fixes array analysis for inplace binary operators.
As title. The `expr.fn` is now the **inplace operator** in the case
of an "inplace_binop" expr node. Adds test in array analysis to
ensure that empty slices hold equivalent (it works correctly without
this change but it may catch future problems). Adds a test to parfors
slice tests that checks the issue, that binary operators were
failing array analysis.

Fixes #3461, Fixes #3554