Rewrite pruned conditionals as their evaluated constants.
This makes it so that when a branch is pruned, the condition
that is rewritten is written to a true/false bit that matches the
evaluated value of the condition. This ensures that an object mode
fall back compilation has IR with equivalent logic.

This bug is also caused by IR mutation being carried over into
object mode. An additional fix will be made for this problem.

Fixes #3879