Prevent mutation of objmode fallback IR.
As title. This prevents the mutation of the IR preserved for
use in the objmode fall back compilation pipeline through the
use of a deepcopy. It also removes an invalid array inlining
behaviour that was occurring in object mode. The new tests hit
the branches in array inlining where no type information is
present, the existing tests hit the branches where type
information is present.

Fixes #2955
Fixes #3239
Fixes #3289
Fixes #3413
Fixes #3659
Fixes #3803