implement realloc and tests
Contrary to the cpython implementation, which is called 'resize', this
one is called 'realloc'. The original 'resize' also changes the 'size'
of the object, which causes hassle when using 'setitem'. Instead,
'realloc' only changes the amount of memory that has been allocated for
future insertions. It is then up to 'setitem' to insert the value
correctly and update the 'size' of the list.