hyperlink repomap
When browsing the repomap I often find myself wanting to click on a file
to view it in github. This implements a custom Sphinx role to insert
exactly those hyperlinks. It has two exceptions, if curly braces are
encountered, e.g. `file.{c, h}`two nodes will be inserted, one for each
file, including a separator between them. If a glob (`*`) is
encountered, a link to the directory is inserted.  Otherwise the text is
linked verbatim. This also has the nice side-effect of checking that all
files mentioned in the repomap really do exist. And in this case
`appveyor.yml` seems to no longer exist so it should probably be
removed. Switching to this role for all files in repomap is also done.