fix incorrect use of memcpy
When using memcpy on overlapping regions of memory, the behaviour is undefined.
For safety, use memmove instead. This largely affected the numba_list_pop
function. As part of this change, the function itself was also refactored to
remove all copied code and now there is a single check to see if the pop was
requested from somewhere other than the end of the list, in which case the
memmove is triggered.