fix raising IndexError
In setitem and getitem, the validity of the index is checked at the
compiler level in `handle_index`. So an invalid index *should* never
propagate to the c-level. Note that, there are currently still guards in
place to ensure we don't have undefined behaviour. But this, also means
such IndexErrors will no longer be raised silently and we will always
get an AssertionError if an invalid index makes it to the c level.