test for list in list and fix memleak
This tests using nested lists. Also, this fixes a memleak in the equals
implementation which was using `zip` to simultaneously iterate over both
lists. Unfortunately `zip` detects the end of an iterable by exception,
which in turn leaks memory in Numba. Using a simple `for i in range`
fixes this.