fix using arrays in lists
This test highlighted two issues, firstly, it wasn't possible to create
lists of array becasue the wrong builder instance was being used and
secondly it wasn't possible to compare lists of arrays since comparing
arrays requires the use of np.any. This commit fixes both to those
issues: the correct builder is used and a special case is introduced for
equals, when arrays are used as item_type in lists.