fix bug in typed-list refinement
This commit does three things:

To begin with, there were actually two severe bugs in `is_prceise`.
First, the negation was missing and secondly the trailing comma meant
that a tuple was being returned. Thus, a list would have always been
deemed to be precise, if there had been any tests for it.

Then, add a simple test to check if refinement works within and outside
of a jit context.

Lastly, After fixing the refinement of the typed-list in a jit context,
it turned out that there is an issue with `overload_method`. The issue
is that, in case of calling an overloaded method on an imprecise
container type, e.g.  typed-list, a function with `overload_method` must
refine the type and return a tuple of the signature that the
implementation should have had and the implementation. While `overload`
does have support for receiving such a tuple, `overload_method` does not
and instead an error regarding a tuple not being callable is raised.
This is fixed now too.