fix type inference for insert
Insert can be used as an alternative to append. This means it can also
be used to insert the first element into a list, when the type of the
list is still undefined. Hence, we need to add code to refine the list
in such cases and give it the type of the item being added. Tests are
included.