Fix bounds checking for slices on shape-0 arrays
In NumPy with a = np.empty((0, 1)), a[0, 0] raises IndexError, but a[:, 0]
does not.

It is impossible to distinguish these two cases in get_item_pointer, which
both want a pointer to the "start" of the empty array. So I have factored out
the bounds checking into a helper function, and called it separately in
basic_indexing() (the same as is already done with wraparound correction).
That way, we turn of bounds checking for slice indices on shape 0 axes. Note
that slice indices are already not bounds checked from above, e.g., a[0:100]
works even if a has fewer than 100 elements. It works this way even on Python
lists.