[WIP] Add `bool` overload for CharSeq and UnicodeCharSeq
The test added in `test_unicode_array` succeeds, but the one in
`test_array_methods` does not.  It raises an error
`NotImplementedError: No definition for lowering <class 'bool'>([unichr x 5],) -> bool`.
I'm not sure why the definition of `bool` that works in the first test is not found
in the later.