Add support for array_like arguments in `argwhere`
Also correct the implementation of `flatnonzero` when taking a "False-y" object that does not
pass `type_can_asarray` (e.g. an empty set or string).  The test is now failing because `bool`
is not defined; see my other PR for that.

The array_like tests will likely fail for older versions of numpy.  I'll make another commit to
fix this (similar to `flatnonzero`) once the full CI suite runs.