fix refining list by using extend on an iterator
The interface for refining on extend for the typed list is a bit
'wobbly'. The type of the underlying thing to extend from, may be
encoded in three different ways. Specifically, certain kinds of
generators or iterators will expose the type of their items using the
`yield_type` attribute. So, we add a test to expose that missing
functionality via a test and add reading the attribute on extend to
enable this functionality.